+++ handler/errors.go
@@ -1,6 +1,11 @@
 package handler
 
-import "fmt"
+import (
+	"fmt"
+
+	"github.com/dogmatiq/dogma"
+	"github.com/dogmatiq/enginekit/message"
+)
 
 // EmptyInstanceIDError indicates that an aggregate or process message handler has
 // attempted to route a message to an instance with an empty ID.
@@ -8,13 +13,15 @@ type EmptyInstanceIDError struct {
 	HandlerName string
 	HandlerKey  string
 	HandlerType Type
+	Message     dogma.Message
 }
 
 func (e EmptyInstanceIDError) Error() string {
 	return fmt.Sprintf(
-		"the '%s' %s message handler attempted to route a message to an empty instance ID",
+		"the '%s' %s message handler attempted to route a %s message to an empty instance ID",
 		e.HandlerName,
 		e.HandlerType,
+		message.TypeOf(e.Message),
 	)
 }
 
@@ -28,7 +35,7 @@ type NilRootError struct {
 
 func (e NilRootError) Error() string {
 	return fmt.Sprintf(
-		"the '%s' %s message handler produced a nil root",
+		"the '%s' %s message handler returned a nil root from New()",
 		e.HandlerName,
 		e.HandlerType,
 	)
@@ -41,6 +48,7 @@ type EventNotRecordedError struct {
 	HandlerKey   string
 	InstanceID   string
 	WasDestroyed bool
+	Message      dogma.Message
 }
 
 func (e EventNotRecordedError) Error() string {
@@ -51,9 +59,10 @@ func (e EventNotRecordedError) Error() string {
 	}
 
 	return fmt.Sprintf(
-		"the '%s' aggregate message handler %s the '%s' instance without recording an event",
+		"the '%s' aggregate message handler %s the '%s' instance without recording an event while handling a %s command",
 		e.HandlerName,
 		s,
 		e.InstanceID,
+		message.TypeOf(e.Message),
 	)
 }