+++ handler/errors_test.go
@@ -1,6 +1,7 @@
 package handler_test
 
 import (
+	"github.com/dogmatiq/enginekit/fixtures"
 	. "github.com/dogmatiq/enginekit/handler"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
@@ -11,11 +12,13 @@ var _ = Describe("type EmptyInstanceIDError", func() {
 		It("returns a meaningful error description", func() {
 			err := EmptyInstanceIDError{
 				HandlerName: "<name>",
+				HandlerKey:  "<key>",
 				HandlerType: AggregateType,
+				Message:     fixtures.MessageA1,
 			}
 
 			Expect(err.Error()).To(Equal(
-				"the '<name>' aggregate message handler attempted to route a message to an empty instance ID",
+				"the '<name>' aggregate message handler attempted to route a fixtures.MessageA message to an empty instance ID",
 			))
 		})
 	})
@@ -26,11 +29,12 @@ var _ = Describe("type NilRootError", func() {
 		It("returns a meaningful error description", func() {
 			err := NilRootError{
 				HandlerName: "<name>",
+				HandlerKey:  "<key>",
 				HandlerType: AggregateType,
 			}
 
 			Expect(err.Error()).To(Equal(
-				"the '<name>' aggregate message handler produced a nil root",
+				"the '<name>' aggregate message handler returned a nil root from New()",
 			))
 		})
 	})
@@ -44,10 +48,11 @@ var _ = Describe("type EventNotRecordedError", func() {
 					HandlerName:  "<name>",
 					InstanceID:   "<instance>",
 					WasDestroyed: false,
+					Message:      fixtures.MessageA1,
 				}
 
 				Expect(err.Error()).To(Equal(
-					"the '<name>' aggregate message handler created the '<instance>' instance without recording an event",
+					"the '<name>' aggregate message handler created the '<instance>' instance without recording an event while handling a fixtures.MessageA command",
 				))
 			})
 		})
@@ -58,10 +63,11 @@ var _ = Describe("type EventNotRecordedError", func() {
 					HandlerName:  "<name>",
 					InstanceID:   "<instance>",
 					WasDestroyed: true,
+					Message:      fixtures.MessageA1,
 				}
 
 				Expect(err.Error()).To(Equal(
-					"the '<name>' aggregate message handler destroyed the '<instance>' instance without recording an event",
+					"the '<name>' aggregate message handler destroyed the '<instance>' instance without recording an event while handling a fixtures.MessageA command",
 				))
 			})
 		})